medguard_project/requirements.txt
# requirements.txt

django
django-cors-headers
pandas
ollama
markdown
pytesseract
Pillow
thefuzz
python-Levenshtein
opencv-python-headless
numpy
google-cloud-vision
dotenv
